no issues after testing

+++ d2l/d2l.py
@@ -199,7 +199,7 @@ def load_data_fashion_mnist(batch_size, resize=None):
 # Defined in file: ./chapter_linear-networks/softmax-regression-scratch.md
 def accuracy(y_hat, y):
     if y_hat.shape[1] > 1:
-        return float((y_hat.argmax(axis=1).astype('float32') == y.astype('float32')).sum())
+        return float((y_hat.argmax(axis=1) == y.astype('float32')).sum())
     else:
         return float((y_hat.astype('int32') == y.astype('int32')).sum())
 
@@ -337,7 +337,7 @@ def download(name, cache_dir='../data'):
 def download_extract(name, folder=None):
     """Download and extract a zip/tar file."""
     fname = download(name)
-    base_dir = os.path.dirname(fname) 
+    base_dir = os.path.dirname(fname)
     data_dir, ext = os.path.splitext(fname)
     if ext == '.zip':
         fp = zipfile.ZipFile(fname, 'r')
@@ -777,9 +777,9 @@ def build_array(lines, vocab, num_steps, is_source):
 def load_data_nmt(batch_size, num_steps, num_examples=1000):
     text = preprocess_nmt(read_data_nmt())
     source, target = tokenize_nmt(text, num_examples)
-    src_vocab = d2l.Vocab(source, min_freq=3, 
+    src_vocab = d2l.Vocab(source, min_freq=3,
                           reserved_tokens=['<pad>', '<bos>', '<eos>'])
-    tgt_vocab = d2l.Vocab(target, min_freq=3, 
+    tgt_vocab = d2l.Vocab(target, min_freq=3,
                           reserved_tokens=['<pad>', '<bos>', '<eos>'])
     src_array, src_valid_len = build_array(
         source, src_vocab, num_steps, True)
@@ -1230,10 +1230,10 @@ def train_gluon_ch11(tr_name, hyperparams, data_iter, num_epochs=2):
 
 
 # Defined in file: ./chapter_computational-performance/hybridize.md
-class benchmark:    
+class benchmark:
     def __init__(self, description = 'Done in %.4f sec'):
         self.description = description
-        
+
     def __enter__(self):
         self.timer = d2l.Timer()
         return self
@@ -1331,7 +1331,7 @@ def train_ch13(net, train_iter, test_iter, loss, trainer, num_epochs,
 
 
 # Defined in file: ./chapter_computer-vision/fine-tuning.md
-d2l.DATA_HUB['hotdog'] = (d2l.DATA_URL+'hotdog.zip', 
+d2l.DATA_HUB['hotdog'] = (d2l.DATA_URL+'hotdog.zip',
                          'fba480ffa8aa7e0febbb511d181409f899b9baa5')
 
 
@@ -1548,7 +1548,7 @@ def reorg_train_valid(data_dir, labels, valid_ratio):
 # Defined in file: ./chapter_computer-vision/kaggle-gluon-cifar10.md
 def reorg_test(data_dir):
     for test_file in os.listdir(data_dir + 'test'):
-        copyfile(data_dir + 'test/' + test_file, 
+        copyfile(data_dir + 'test/' + test_file,
                  data_dir + 'train_valid_test/test/unknown/')
 
 
@@ -1558,7 +1558,7 @@ d2l.DATA_HUB['dog_tiny'] = (d2l.DATA_URL + 'kaggle_dog_tiny.zip',
 
 
 # Defined in file: ./chapter_natural-language-processing-pretraining/word2vec-dataset.md
-d2l.DATA_HUB['ptb'] = (d2l.DATA_URL + 'ptb.zip', 
+d2l.DATA_HUB['ptb'] = (d2l.DATA_URL + 'ptb.zip',
                        '319d85e578af0cdc590547f26231e4e31cdf1e42')
 
 
@@ -2080,7 +2080,7 @@ def read_snli(data_dir, is_train):
     """Read the SNLI dataset into premises, hypotheses, and labels."""
     def extract_text(s):
         # Remove information that will not be used by us
-        s = re.sub('\(', '', s) 
+        s = re.sub('\(', '', s)
         s = re.sub('\)', '', s)
         # Substitute two or more consecutive whitespace with space
         s = re.sub('\s{2,}', ' ', s)
@@ -2114,7 +2114,7 @@ class SNLIDataset(gluon.data.Dataset):
         print('read ' + str(len(self.premises)) + ' examples')
 
     def pad(self, lines):
-        return np.array([d2l.trim_pad(self.vocab[line], self.num_steps, 
+        return np.array([d2l.trim_pad(self.vocab[line], self.num_steps,
                                       self.vocab['<pad>']) for line in lines])
 
     def __getitem__(self, idx):
